add: using clerk middleware in album routes,
also added a test to check if saving an album to the collection works and if the album is associated with the correct user, and updated the frontend to fetch albums from the new collection route instead of fetching albums by user id

+++ routes/album.js
@@ -1,7 +1,22 @@
 const express = require("express");
 const { getSpotifyAccessToken } = require("./utils/spotify");
 const Album = require("../models/Albums");
+const { getAuth } = require("@clerk/express");
 const router = express.Router();
+
+function ensureAuthenticated(req, res, next) {
+    const { userId } = getAuth(req);
+
+    if (!userId) {
+        return res.status(401).json({ error: "Unauthorized" });
+    }
+
+    req.userId = userId;
+    next();
+}
+
+
+
 
 router.get("/album/:id", async(req, res) =>
 {
@@ -36,10 +51,10 @@ router.get("/album/:id", async(req, res) =>
     }
 });
 
-router.post("/album", async(req, res) => 
+router.post("/album", ensureAuthenticated, async(req, res) => 
 {
     try {
-        const album = await Album.create(req.body);
+        const album = await Album.create({ ...req.body, userId: req.userId });
         res.status(201).json(album);
     } catch (error) {
         console.log(error);
@@ -47,10 +62,12 @@ router.post("/album", async(req, res) =>
     }
 });
 
-router.get("/user/:userId", async(req, res) =>
+router.get("/collection", ensureAuthenticated, async(req, res) =>
 {
+    const userId = req.userId;
+
     try {
-        const albums = await Album.find({ userId: req.params.userId });
+        const albums = await Album.find({ userId });
         res.json(albums);
     } catch (error) {
         console.log(error);
@@ -58,9 +75,9 @@ router.get("/user/:userId", async(req, res) =>
     }
 });
 
-router.delete("/album/:id", async(req, res) => {
+router.delete("/album/:id", ensureAuthenticated, async(req, res) => {
     try {
-        await Album.findOneAndDelete({ spotifyId: req.params.id });
+        await Album.findOneAndDelete({ spotifyId: req.params.id, userId: req.userId });
         res.json({ message: "Album removed from collection" });
     } catch (error) {
         console.log(error)
